requirements.txt
organize-tool>=2.4.0
pyyaml>=6.0
mutagen>=1.45.0
Pillow>=9.0.0
